Potential bug found in Transpose inplace
